Move swift modules into subdirectory
