+++ compiler/core/bsconfig.json
@@ -3,7 +3,12 @@
 {
   "name": "lona-compiler-core",
   "version": "0.1.0",
-  "sources": ["src"],
+  "sources": [
+    {
+      "dir": "src",
+      "subdirs": true
+    }
+  ],
   "package-specs": {
     "module": "commonjs",
     "in-source": true
