+++ compiler/core/src/swift/swiftComponent.bs.js
@@ -7,10 +7,10 @@ var Curry                          = require("bs-platform/lib/js/curry.js");
 var Caml_obj                       = require("bs-platform/lib/js/caml_obj.js");
 var Pervasives                     = require("bs-platform/lib/js/pervasives.js");
 var LodashUpperfirst               = require("lodash.upperfirst");
-var Layer$LonaCompilerCore         = require("./layer.bs.js");
-var Logic$LonaCompilerCore         = require("./logic.bs.js");
-var Decode$LonaCompilerCore        = require("./decode.bs.js");
-var StringMap$LonaCompilerCore     = require("./stringMap.bs.js");
+var Layer$LonaCompilerCore         = require("../layer.bs.js");
+var Logic$LonaCompilerCore         = require("../logic.bs.js");
+var Decode$LonaCompilerCore        = require("../decode.bs.js");
+var StringMap$LonaCompilerCore     = require("../stringMap.bs.js");
 var SwiftLogic$LonaCompilerCore    = require("./swiftLogic.bs.js");
 var SwiftFormat$LonaCompilerCore   = require("./swiftFormat.bs.js");
 var SwiftDocument$LonaCompilerCore = require("./swiftDocument.bs.js");
